requirements.txt
kafka-python
requests
textblob

python -m textblob.download_corpora
